Rename files to have a leading zero in day number
