+++ DonKarlssonSan-JavaScript/day02-checksum.js
