+++ DonKarlssonSan-JavaScript/day03.js
