+++ DonKarlssonSan-JavaScript/day04-passphrase-validator.js
