+++ DonKarlssonSan-JavaScript/day06.js
