+++ DonKarlssonSan-JavaScript/day08-input.txt
