+++ DonKarlssonSan-JavaScript/day09.js
